Implemented BrowserWindow.Maximized
Implemented BrowserWindow.Maximized as per documentation at
https://msdn.microsoft.com/en-us/library/microsoft.visualstudio.testtools.uitesting.applicationundertest.maximized.aspx

Also added the Properties Width and Height on BrowserWindow. Added a
test for it. The test is passing.

Selected the default values to be non standard so it would be easy to
assert that the BrowserWindow is maximized, without getting the PCs
width and height.