Implemented HtmlListItem
Implemented HtmlListItem based on the documentation at
https://msdn.microsoft.com/en-us/library/microsoft.visualstudio.testtools.uitesting.htmlcontrols.htmllistitem.aspx

Based on the documentation, it has 2 public properties that are not
Inherited: DisplayText and Selected. And one method that's unique to
this class: Select(), marks this HtmlListItem as selected in the parent
container. Implemented this as simple as possible. Added a basic test
for all the fields.